[dice] move dice lib
The DICE lib does not just generate certificates, it also generates
keys. Therefore, it is more appropriately placed as a generic
silicon_creator library.
